chore(ci): switch new-src/ to src/ and new-test/ to test/
