+++ src/is-promise.ts
