+++ test/edge-cases.test.ts
@@ -1,33 +1,44 @@
-import { expect, test } from "bun:test";
+import { readFile } from "node:fs/promises";
 import * as v from "valibot";
+import { expect, test } from "vitest";
 import z from "zod";
-import { zagora } from "../src/index.ts";
+import { isInternalError } from "../src/errors";
+import { zagora } from "../src/index";
 
-test("typed errors handler arg should be second arg when no input schema ", () => {
+test("typed errors should be in options - when input schema is defined", () => {
   const errorSchemas = {
-    SOME_ERR: z.object({
-      type: z.literal("SOME_ERR"),
-      msg: z.string().default("Unknown error"),
+    BAR_ERR: z.object({
+      msg: z.string(),
       foo: z.number().min(100).max(999).default(500),
     }),
   };
 
   const func = zagora()
     .errors(errorSchemas)
-    .handler((_, err) => {
-      throw err.SOME_ERR({ msg: "Custom error" });
-    });
+    // .errors({ FOO: z.object({ msg: z.string() }) })
+    .handler(({ errors }) => {
+      throw errors.BAR_ERR({ msg: "Custom error" });
+
+      // NOTE: simulate return, because the return types break
+      // when no return and no `input` schema
+
+      // biome-ignore lint/correctness/noUnreachable: bruh
+      return "foo";
+    })
+    .callable();
 
   const res = func();
-  expect(res.isDefined).toBe(true);
+  expect(res.ok).toBe(false);
 
-  if (res.isDefined && res.error.type === "SOME_ERR") {
-    expect(Object.keys(errorSchemas)[0]).toBe(res.error.type);
-    expect(res.error.type).toBe("SOME_ERR");
-    expect(res.error.msg).toBe("Custom error");
-    expect(res.error.foo).toBe(500);
+  if (res.error) {
+    expect(Object.keys(errorSchemas)[0]).toBe(res.error.kind);
+    expect(res.error.kind).toBe("BAR_ERR");
+    if (res.error.kind === "BAR_ERR") {
+      expect(res.error.msg).toBe("Custom error");
+      expect(res.error.foo).toBe(500);
+    }
   } else {
-    throw new Error("Expected SOME_ERR, but got: " + JSON.stringify(res.error));
+    throw new Error(`Expected BAR_ERR, but got: ${JSON.stringify(res.error)}`);
   }
 });
 
@@ -41,14 +52,17 @@ test("Zod tuple with default values - basic case", async () => {
         foo: z.string().min(1),
       }),
     )
-    .handler((speed, retry) => {
+    .handler((_, speed, retry) => {
       return { foo: `${speed}-${retry}` };
-    });
+    })
+    .callable();
 
-  const [resHello, errHello] = hello("fast");
+  const res = hello("fast");
 
-  expect(errHello).toBe(null);
-  expect(resHello).toEqual({ foo: "fast-123" });
+  expect(res.ok).toBe(true);
+  if (res.ok) {
+    expect(res.data).toEqual({ foo: "fast-123" });
+  }
 });
 
 test("Zod tuple with required arg, default value arg, and optional arg", async () => {
@@ -63,14 +77,17 @@ test("Zod tuple with required arg, default value arg, and optional arg", async (
         foo: z.string().min(1),
       }),
     )
-    .handler((speed, retry, str) => {
+    .handler(async (_, speed, retry, str) => {
       return { foo: `${speed}-${retry}-${str}` };
-    });
+    })
+    .callable();
 
-  const [resHello, errHello] = hello("fast");
+  const res = await hello("fast");
 
-  expect(errHello).toBe(null);
-  expect(resHello).toEqual({ foo: "fast-123-undefined" });
+  expect(res.ok).toBe(true);
+  if (res.ok) {
+    expect(res.data).toEqual({ foo: "fast-123-undefined" });
+  }
 });
 
 test("Zod tuple with default values - both args provided", async () => {
@@ -83,14 +100,17 @@ test("Zod tuple with default values - both args provided", async () => {
         foo: z.string().min(1),
       }),
     )
-    .handler(async (speed, retry) => {
+    .handler(async (_, speed, retry) => {
       return { foo: `${speed}-${retry}` };
-    });
+    })
+    .callable();
 
-  const [resHello, errHello] = await hello("slow", 456);
+  const res = await hello("slow", 456);
 
-  expect(errHello).toBe(null);
-  expect(resHello).toEqual({ foo: "slow-456" });
+  expect(res.ok).toBe(true);
+  if (res.ok) {
+    expect(res.data).toEqual({ foo: "slow-456" });
+  }
 
   const hello2 = zagora()
     .input(z.tuple([SpeedSchema, z.number().default(123)]))
@@ -101,14 +121,17 @@ test("Zod tuple with default values - both args provided", async () => {
         })
         .strict(),
     )
-    .handler(async (speed, retry) => {
+    .handler(async (_, speed, retry) => {
       return { foo: `${speed}-${retry}`, bar: "barry" };
-    });
+    })
+    .callable();
 
-  const [resHello2, errHello2] = await hello2("slow", 456);
+  const failingHello = await hello2("slow", 456);
 
-  expect(errHello2?.reason).toContain("utput validation failed");
-  expect(resHello2).toBeEmpty();
+  expect(failingHello.ok).toBe(false);
+  if (!failingHello.ok) {
+    expect(failingHello.error.message).toContain("Output validation failed");
+  }
 });
 
 test("Zod tuple with multiple defaults", async () => {
@@ -127,14 +150,15 @@ test("Zod tuple with multiple defaults", async () => {
         foo: z.string().min(1),
       }),
     )
-    .handler(async (speed, retry, extra) => {
+    .handler(async (_, speed, retry, extra) => {
       return { foo: `${speed}-${retry}-${extra}` };
-    });
+    })
+    .callable();
 
-  const [resHello, errHello] = await hello("normal");
+  const res = await hello("normal");
 
-  expect(errHello).toBe(null);
-  expect(resHello).toEqual({ foo: "normal-123-bruh" });
+  expect(res.ok).toBe(true);
+  expect((res as any).data).toEqual({ foo: "normal-123-bruh" });
 });
 
 test("Valibot tuple with optional and default value", async () => {
@@ -147,14 +171,15 @@ test("Valibot tuple with optional and default value", async () => {
         foo: v.pipe(v.string(), v.minLength(1)),
       }),
     )
-    .handler(async (speed, retry) => {
+    .handler(async (_, speed, retry) => {
       return { foo: `${speed}-${retry}` };
-    });
+    })
+    .callable();
 
-  const [resHello, errHello] = await hello("fast");
+  const res = await hello("fast");
 
-  expect(errHello).toBe(null);
-  expect(resHello).toEqual({ foo: "fast-123" });
+  expect(res.ok).toBe(true);
+  expect((res as any).data).toEqual({ foo: "fast-123" });
 });
 
 test("Valibot tuple with optional without default", async () => {
@@ -167,19 +192,20 @@ test("Valibot tuple with optional without default", async () => {
         foo: v.pipe(v.string(), v.minLength(1)),
       }),
     )
-    .handler(async (speed, retry) => {
+    .handler(async (_, speed, retry) => {
       return { foo: `${speed}-${retry}` };
-    });
+    })
+    .callable();
 
-  const [resHello, errHello] = await hello("slow");
+  const res = await hello("slow");
 
-  expect(errHello).toBe(null);
-  expect(resHello).toEqual({ foo: "slow-undefined" });
+  expect(res.ok).toBe(true);
+  expect((res as any).data).toEqual({ foo: "slow-undefined" });
 
-  const [resHello2, errHello2] = await hello("slow", 222);
+  const bruh = await hello("slow", 222);
 
-  expect(errHello2).toBe(null);
-  expect(resHello2).toEqual({ foo: "slow-222" });
+  expect(bruh.ok).toBe(true);
+  expect((bruh as any).data).toEqual({ foo: "slow-222" });
 });
 
 test("Valibot tuple with default values - both args provided", async () => {
@@ -198,14 +224,15 @@ test("Valibot tuple with default values - both args provided", async () => {
         foo: v.pipe(v.string(), v.minLength(1)),
       }),
     )
-    .handler(async (speed, retry, str) => {
+    .handler(async (_, speed, retry, str) => {
       return { foo: `${speed}-${retry}-${str}` };
-    });
+    })
+    .callable();
 
-  const [resHello, errHello] = await hello("slow", 456);
+  const res = await hello("slow", 456);
 
-  expect(errHello).toBe(null);
-  expect(resHello).toEqual({ foo: "slow-456-str" });
+  expect(res.ok).toBe(true);
+  expect((res as any).data).toEqual({ foo: "slow-456-str" });
 });
 
 test("Tuple without defaults - all args required - zod", async () => {
@@ -218,14 +245,15 @@ test("Tuple without defaults - all args required - zod", async () => {
         foo: z.string().min(1),
       }),
     )
-    .handler(async (speed, retry) => {
+    .handler(async (_, speed, retry) => {
       return { foo: `${speed}-${retry}` };
-    });
+    })
+    .callable();
 
-  const [resHello, errHello] = await hello("fast", 456);
+  const res = await hello("fast", 456);
 
-  expect(errHello).toBe(null);
-  expect(resHello).toEqual({ foo: "fast-456" });
+  expect(res.ok).toBe(true);
+  expect((res as any).data).toEqual({ foo: "fast-456" });
 });
 
 test("Tuple without defaults - all args required - valibot", async () => {
@@ -238,22 +266,24 @@ test("Tuple without defaults - all args required - valibot", async () => {
         foo: v.pipe(v.string(), v.minLength(1)),
       }),
     )
-    .handler(async (speed, retry) => {
+    .handler(async (_, speed, retry) => {
       return { foo: `${speed}-${retry}` };
-    });
+    })
+    .callable();
 
-  const [resHello, errHello] = await hello("fast", "sasa");
+  const res = await hello("fast", "sasa");
 
-  expect(errHello).toBe(null);
-  expect(resHello).toEqual({ foo: "fast-sasa" });
+  expect(res.ok).toBe(true);
+  expect((res as any).data).toEqual({ foo: "fast-sasa" });
 
-  const [_, errHello2] = await hello("fast");
+  // @ts-expect-error - should error because missing second required arg
+  const result = await hello("fast");
 
-  expect(errHello2).not.toBeNull();
-  expect(errHello2?.reason).toContain("Input validation failed");
+  expect(result.ok).toBe(false);
+  expect(result.error?.message).toContain("Input validation failed");
 });
 
-test("Tuple without defaults - missing required arg should fail", async () => {
+test("Zod Tuple without defaults - missing required arg should fail", async () => {
   const SpeedSchema = z.enum(["slow", "normal", "fast"]);
 
   const hello = zagora()
@@ -263,13 +293,189 @@ test("Tuple without defaults - missing required arg should fail", async () => {
         foo: z.string().min(1),
       }),
     )
-    .handler(async (speed, retry) => {
+    .handler(async (_, speed, retry) => {
       return { foo: `${speed}-${retry}` };
-    });
+    })
+    .callable();
 
-  const [resHello, errHello] = await hello("fast");
+  // @ts-expect-error - should error because missing second required arg
+  const res = await hello("fast");
 
-  expect(resHello).toBe(null);
-  expect(errHello).toBeDefined();
-  expect(errHello?.message).toContain("Invalid input");
+  expect(res.ok).toBe(false);
+  expect(res.error?.message).toContain("Input validation failed");
+});
+
+test("Handler without input schema should work", () => {
+  const errorSchemas = {
+    BARRY_ERR: z.object({
+      msg: z.string(),
+      foo: z.number().min(100).max(999).default(500),
+    }),
+  };
+
+  const func = zagora()
+    // no .input() call
+    .errors(errorSchemas)
+    .handler(({ errors }) => {
+      throw errors.BARRY_ERR({ msg: "Some custom error" });
+      // biome-ignore lint/correctness/noUnreachable: bruh
+      return "foo";
+    })
+    .callable();
+
+  const res = func();
+  expect(res.ok).toBe(false);
+
+  if (res.error) {
+    expect(Object.keys(errorSchemas)[0]).toBe(res.error.kind);
+    expect(res.error.kind).toBe("BARRY_ERR");
+    if (res.error.kind === "BARRY_ERR") {
+      expect(res.error.msg).toBe("Some custom error");
+      expect(res.error.foo).toBe(500);
+    }
+  } else {
+    throw new Error(
+      `Expected BARRY_ERR, but got: ${JSON.stringify(res.error)}`,
+    );
+  }
+});
+
+test("Handler without input schema and no errors should work", () => {
+  const func = zagora({ autoCallable: true }).handler(({ context }) => {
+    return { result: "success", ctx: context };
+  });
+
+  const res = func();
+
+  expect(res.ok).toBe(true);
+  if (res.ok) {
+    expect(res.data).toEqual({ result: "success", ctx: undefined });
+  }
+});
+
+test("wrapping external async functions in pseudo-sync `.handler` fn", async () => {
+  // Basic async function that can succeed or fail
+  const basicAsyncFn = async (input: string): Promise<string> => {
+    if (input !== "foobie") {
+      throw new Error("Basic async failed");
+    }
+    return `processed: ${input}`;
+  };
+
+  const basicWrapper = zagora()
+    .handler((_) => basicAsyncFn("foobie"))
+    .callable();
+
+  const res1 = await basicWrapper();
+  expect(res1.ok).toBe(true);
+  expect((res1 as any).data).toBe("processed: foobie");
+
+  // Test basic wrapper without schemas - failure
+  const failingBasicWrapper = zagora()
+    .handler((_) => basicAsyncFn("fail one"))
+    .callable();
+
+  const res2 = await failingBasicWrapper();
+  expect(res2.ok).toBe(false);
+
+  if (!res2.ok && isInternalError(res2.error)) {
+    expect(res2.error.message).toContain("Async handler threw");
+    expect(res2.error.cause).toBeInstanceOf(Error);
+    expect((res2.error.cause as Error).message).toBe("Basic async failed");
+  }
+
+  // Test with non-existent file
+  const fsWrapper = zagora()
+    .input(z.string())
+    .output(z.string())
+    .handler((_, filepath) => readFile(filepath, "utf-8"))
+    .callable();
+
+  const res3 = await fsWrapper("non-existent-file.txt");
+
+  expect(res3.ok).toBe(false);
+  if (!res3.ok && isInternalError(res3.error)) {
+    expect(res3.error.message).toContain("Async handler threw");
+    expect(res3.error.cause).toBeInstanceOf(Error);
+    expect((res3.error.cause as Error).message).toContain("ENOENT");
+  } else {
+    expect(
+      false,
+      "Should throw internal error cought from 'pseudo-sync' handler returning promise",
+    ).toBeInstanceOf(true);
+  }
+
+  // Test with package.json (should exist)
+  const res4 = await fsWrapper("package.json");
+  expect(res4.ok).toBe(true);
+
+  if (res4.ok) {
+    expect(typeof res4.data).toBe("string");
+    expect(res4.data).toContain('name":');
+    expect(res4.data).toContain('"zagora');
+  } else {
+    expect(
+      false,
+      "Should not throw error caught when 'pseudo-sync' handler returns promise",
+    ).toBeInstanceOf(true);
+  }
+});
+
+test("wrapping sync throwing functions (JSON.parse -> safeJsonParse) in .handler", async () => {
+  const someJson = `{"name": "zagora"}`;
+  const safeJsonParse = zagora()
+    .input(z.string())
+    .output(z.object({ name: z.string() }))
+    .handler((_, input) => JSON.parse(input))
+    .callable();
+
+  const res = safeJsonParse(someJson);
+  expect(res.ok).toBe(true);
+
+  if (res.ok) {
+    expect(res.data.name).toBe("zagora");
+  } else {
+    expect(false, "Should not have error").toBeInstanceOf(true);
+  }
+
+  const res2 = safeJsonParse(`foo": 123`);
+  expect(res2.ok).toBe(false);
+
+  if (res2.error && isInternalError(res2.error)) {
+    expect(res2.error.kind).toBe("UNKNOWN_ERROR");
+    expect(res2.error.message).toContain("Sync handler threw unknown err");
+    expect((res2.error.cause as Error).message).toContain("Unexpected");
+  } else {
+    expect(false, "Should throw error").toBeInstanceOf(true);
+  }
+});
+
+test("proper return type for sync and async handlers", async () => {
+  const someJson = `{"name": "zagora"}`;
+  const safeAsyncParse = zagora()
+    .handler(async () => JSON.parse(someJson))
+    .callable();
+  const res1 = await safeAsyncParse();
+  expect(res1.ok).toBe(true);
+  expect((res1 as any).data.name).toBe("zagora");
+  const res2 = safeAsyncParse();
+  expect(res2).toBeInstanceOf(Promise);
+  expect(((await res2) as any).data.name).toBe("zagora");
+
+  const safeParseSync = zagora()
+    .handler(() => JSON.parse(someJson))
+    .callable();
+  const res3 = safeParseSync();
+  expect(res3.ok).toBe(true);
+  expect((res3 as any).data.name).toBe("zagora");
+
+  const safeParsePromiseSync = zagora()
+    .handler(() => Promise.resolve(JSON.parse(someJson)))
+    .callable();
+  const res4 = await safeParsePromiseSync();
+  expect(res4.ok).toBe(true);
+  expect((res4 as any).data.name).toBe("zagora");
+  const res5 = safeAsyncParse();
+  expect(res5).toBeInstanceOf(Promise);
+  expect(((await res5) as any).data.name).toBe("zagora");
 });
