+++ vitest.config.ts
@@ -2,8 +2,6 @@ import isCI from "is-ci";
 import { defineConfig } from "vitest/config";
 export default defineConfig({
   test: {
-    include: ["**/new-test/**/*.test.ts", "**/new-test/*.test.ts"],
-    exclude: ["**/node_modules/**"],
     coverage: {
       enabled: process.env.COV === "1" || isCI,
       clean: true,
